Delete previous versions of grid terrain.
