--- construction-engine/Assets/Materials/Resources/Palette.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e136ea784d883924abadc7886031ee29
-folderAsset: yes
-timeCreated: 1521418065
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
