--- construction-engine/Assets/Materials/Resources/Terrain/GridTerrainMaterial.mat.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: a35e9a2fd72650e42bf6bdc65534d122
-timeCreated: 1522606228
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
