--- construction-engine/Assets/Prefabs/Resources/cursor.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 0333f67e77c61ce43aaae5ddb4970268
-timeCreated: 1517702509
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
