--- construction-engine/Assets/Scenes/terrain01.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 30b706231b724e244af62ba2a3b49a3f
-timeCreated: 1517459883
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
