--- construction-engine/Assets/Scenes/terrain02.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c3aafd088b471b84eaa8fe4be2ce33f8
-timeCreated: 1517498874
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
