--- construction-engine/Assets/Scripts/GridTerrain/CheckerboardTerrain.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c2160cc3f9277d644aaa2445a702e811
-timeCreated: 1517766317
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
