--- construction-engine/Assets/Scripts/GridTerrain/GridTerrainArgs.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 14614c3a9e9136e46904856369512fb6
-timeCreated: 1522619278
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
