+++ construction-engine/Assets/Scripts/GridTerrain/GridTerrainData.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 16fea44c87a8cef4cbde760d8f7f55c8
-timeCreated: 1517766317
+guid: 707aa86a9b5a7ab4ab9de7b74fbe1bba
+timeCreated: 1523132412
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
