--- construction-engine/Assets/Scripts/GridTerrain/GridTerrainData2.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 924941de717086a4c9cee732888013bf
-timeCreated: 1522632611
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
