--- construction-engine/Assets/Scripts/GridTerrain/GridTerrainData3.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: ba195cc73528c0c468cdee456060f45f
-timeCreated: 1523117634
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
