--- construction-engine/Assets/Scripts/GridTerrain/GridTerrainData4.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 707aa86a9b5a7ab4ab9de7b74fbe1bba
-timeCreated: 1523132412
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
