--- construction-engine/Assets/Scripts/GridTerrain/SelectableTerrain.cs
@@ -1,12 +1,0 @@
-﻿using Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace GridTerrain
-{
-    public class SelectableTerrain : Selectable
-    {
-    }
-}
