--- construction-engine/Assets/Terrains.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e357322248795fc439e4a6fb5a219f03
-folderAsset: yes
-timeCreated: 1517093998
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
