--- construction-engine/Assets/Terrains/Shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 6a131a401ddbc2b4aa99fadcb31320c0
-folderAsset: yes
-timeCreated: 1522618401
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
