--- construction-engine/Assets/Terrains/Shader/Custom-GridTerrain-AddPass.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: f7f3cc626e2ea274db215ce93a133ea4
-timeCreated: 1522617884
-licenseType: Free
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
