--- construction-engine/Assets/Terrains/Shader/Custom-GridTerrain-FirstPass.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 147cc6316603db540bb1af67ff5ab832
-timeCreated: 1522617882
-licenseType: Free
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
