--- construction-engine/Assets/Terrains/Shader/Custom-TerrainSplatmapCommon.cginc.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 39d97634b5de5404cab4a58cbc1fe9af
-timeCreated: 1522608062
-licenseType: Free
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
