--- construction-engine/Assets/Terrains/Terrain.asset.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 46d12a8ebcd18944f9b5ab22bd11346d
-timeCreated: 1517459898
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
